Remove acceptance test code (#1)
https://trello.com/c/9BppMgM8

This started off with us spotting that this repo contained a password. Even though encrypted the same file contained the code necessary to decrypt the password.

So we needed to remove it and replace it with a different mechanism for providing it. But as you can see, this was just a Java file (and the compiled class which is useless). There is no mechanism for working with the code, especially when you take into account it is hardcoded to just work on one individuals machine.

TL;DR we scrapping this code here (it shouldn't even be here. This is a repo for JMeter files). Instead we have created [sroc-tcm-acceptance-tests](https://github.com/DEFRA/sroc-tcm-acceptance-tests) which is the new home for this file.