Update project to meet Defra standards (#2)
https://github.com/DEFRA/dst-guides/blob/master/github/new_projects.md#new-projects

There is a minimum standard we expect of any repo on GitHub, let alone one that is public. This change covers everything needed to bring it up to that standard.

It won't cover whether the JMeter files themselves need changes. For example, manual versioning appears to have been used which is unnecessary and has resulted in duplication.

* Add license

All our projects should have a license. This is especially important for public repos because without one it means people are not actually permitted access to the code.

* Remove temporary output files

This includes adding a .gitignore to ensure they don't get accidentally added again.

As temporary files created for specific test runs they should not be committed to the source code for the project.

* Update the README

Includes our standard sections and at least now some context for what the project is about.